refactor(views): refatorando as views de contas
